requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
av==15.0.0
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
dashscope==1.24.2
distro==1.9.0
eval_type_backport==0.2.2
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.1
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
openai==1.101.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qwen-agent==0.0.29
qwen-vl-utils==0.0.11
referencing==0.36.2
regex==2025.7.33
requests==2.32.5
rpds-py==0.27.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wheel==0.45.1
yarl==1.20.1
